Add Clash.Normalize.Transformations.Inline
All transformations concerning inlining now live in a separate module.
This is part of work to make the normalization modules quicker to compile.
